Optimize use of RAM, plus leverage INPUT_PULLUP
Moved some strings from RAM to program storage to conserve RAM, plus changed the polarity of the reset and select switches and set the input mode to INPUT_PULLUP to leverage the builtin pull-up resistors in the Arduino and reduce components and simplify the circuit. Reset and select switches are now active low, and the pull-up resistor is internal to the Arduino. Also changed the code to always auto mount AUTORUN.ATR if it exists on startup, mostly eliminating the need for a separate reset switch, the Arduino reset button can now server this purpose, allowing reset and auto mount of AUTORUN.ATR to coexist with a selector button to change the disk. This allows both programatic and manual control of the mounted disk easily.